backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
pydantic==2.10.3
python-multipart==0.0.12
joblib==1.4.2
google-generativeai>=0.8.0
python-dotenv==1.0.1
tavily-python>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
rouge-score>=0.1.2
torch>=2.0.0
